+++ zucchini-ui-e2e-tests/cypress/integration/testRuns.spec.js
@@ -20,53 +20,16 @@ describe("Test runs", () => {
   });
 
   it("should create a test run", () => {
-    cy.server();
-    cy.route("POST", "/api/testRuns/*/import**").as("importCucumberReport");
-
-    cy.log("Créer un tir");
-
-    cy
-      .createTestRun({
-        type: "Create test run",
-        environment: "PREP",
-        name: "Exemple"
-      })
-      .its("url")
-      .then(testRunUrl => {
-        cy.visit(testRunUrl);
-      });
-
-    cy.get("h1").contains(/Tir du [0-9]{2}\/[0-9]{2}\/[0-9]{2}/);
-
-    cy.log("Importer un fichier de résultats Cucumber");
-
-    cy.get("button").contains("Importer").click();
+    cy.get("button").contains("Créer un tir").click();
 
     cy.get("[role=dialog]").within(() => {
-      // Workaround : cucumber-report.json not loaded as base64, even if asked. So... extention change
-      cy.fixture("cucumber-report.json.bin", "base64").then(fileContent => {
-        cy.get("input#file").upload(
-          {
-            fileContent,
-            fileName: "report.json",
-            mimeType: "application/json",
-            encoding: "base64"
-          },
-          {
-            subjectType: "input"
-          }
-        );
-      });
+      cy.get("input#type").type("TYPE");
+      cy.get("input#environment").type("ENV");
+      cy.get("input#name").type("NAME");
 
-      cy.get("button").contains("Importer").click();
+      cy.get("button").contains("Créer").click();
     });
 
-    cy.log("Validate upload with table size");
-
-    cy.wait("@importCucumberReport");
-
-    cy.get("table").within(() => {
-      cy.get("tbody").find("tr").should("have.length.greaterThan", 0);
-    });
+    cy.url().should("include", "/ui/test-runs/");
   });
 });