+++ CHANGELOG.md
@@ -11,6 +11,8 @@ Version 0.0.5-SNAPSHOT
 * Migration to `react-router-dom` (latest version of React Router)
 * Added two new columns to identify the type and the environment of the test run
 * Added a new filter by environment
+* E2E tests with Cypress
+*** Required Node version: 10**
 
 Version 0.0.4-SNAPSHOT
 ----------------------